requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
asyncio-extras==1.3.2
asyncpraw==7.6.1
asyncprawcore==2.3.0
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
discord.py==2.1.0
frozendict==2.3.4
frozenlist==1.3.3
future==0.18.3
gazpacho==1.1
html5lib==1.1
idna==3.4
instagrapi==1.16.30
lxml==4.9.2
multidict==6.0.4
multitasking==0.0.11
newsapi-python==0.2.6
numpy==1.24.1
oauthlib==3.2.2
obliquestrategies==0.3
paginator.py==0.9.4
pandas==1.5.3
Pillow==9.4.0
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.10.2
PySocks==1.7.1
python-dateutil==2.8.2
PyTumblr==0.1.2
pytz==2022.7.1
quote==2.0.4
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
tinydb==4.7.1
tweepy==4.12.1
typing_extensions==4.4.0
update-checker==0.18.0
urllib3==1.26.14
webencodings==0.5.1
yarl==1.8.2
yfinance==0.2.5
